# docs/swagger/swagger.yaml
basePath: /v1
consumes:
    - application/json
definitions:
...
    Member:
        properties:
            created_at:
                description: The timestamp of when the member was created.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: CreatedAt
            deleted_at:
                description: The timestamp of when the member was deleted.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: DeletedAt
            group_uuid:
                description: The UUID of the group.
                example: '"f3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"'
                type: string
                x-go-name: GroupUUID
            id:
                description: The ID of the member.
                example: 1
                format: uint64
                type: integer
                x-go-name: ID
            role:
                description: The role of the member.
                example: '"admin"'
                type: string
                x-go-name: Role
            updated_at:
                description: The timestamp of when the member was last updated.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: UpdatedAt
            user_uuid:
                description: The UUID of the user.
                example: '"f3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"'
                type: string
                x-go-name: UserUUID
            uuid:
                description: The UUID of the member.
                example: '"f3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"'
                type: string
                x-go-name: UUID
        required:
            - id
            - uuid
            - group_uuid
            - user_uuid
            - role
        title: Member represents a member in the system.
        type: object
        x-go-package: github.com/ryo-arima/locky/pkg/entity/response
    MemberRequest:
        properties:
            created_at:
                description: The timestamp of when the member was created.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: CreatedAt
            deleted_at:
                description: The timestamp of when the member was deleted.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: DeletedAt
            group_uuid:
                description: The UUID of the group.
                example: '"f3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"'
                type: string
                x-go-name: GroupUUID
            id:
                description: The ID of the member.
                example: 1
                format: uint64
                type: integer
                x-go-name: ID
            role:
                description: The role of the member.
                example: '"admin"'
                type: string
                x-go-name: Role
            updated_at:
                description: The timestamp of when the member was last updated.
                example: '"2023-01-01T00:00:00Z"'
                format: date-time
                type: string
                x-go-name: UpdatedAt
            user_uuid:
                description: The UUID of the user.
                example: '"f3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"'
                type: string
                x-go-name: UserUUID
            uuid:
                description: The UUID of the member.
                example: '"f3b3b3b3-3b3b-3b3b-3b3b-3b3b3b3b3b3b"'
                type: string
                x-go-name: UUID
        required:
            - group_uuid
            - user_uuid
            - role
        title: MemberRequest represents the request body for member-related operations.
        type: object
        x-go-package: github.com/ryo-arima/locky/pkg/entity/request
    MemberResponse:
        properties:
            code:
                description: The response code.
                example: '"SUCCESS"'
                type: string
                x-go-name: Code
            members:
                description: The list of members.
                items:
                    $ref: '#/definitions/Member'
                type: array
                x-go-name: Members
            message:
                description: The response message.
                example: '"Members retrieved successfully"'
                type: string
                x-go-name: Message
        required:
            - code
            - message
            - members
        title: MemberResponse represents the response body for member-related operations.
        type: object
        x-go-package: github.com/ryo-arima/locky/pkg/entity/response
...
host: localhost:8080
info:
    description: '# This is a sample Locky API'
    title: Locky API.
    version: 1.0.0
paths:
...
    /internal/members:
        get:
            description: Get a list of all members in the system.
            operationId: getMembersInternal
            responses:
                "200":
                    description: A list of members.
                    schema:
                        $ref: '#/definitions/MemberResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/MemberResponse'
            summary: Get a list of members.
            tags:
                - members
        post:
            description: Create a new member with the provided information.
            operationId: createMemberInternal
            parameters:
                - description: The member to create.
                  in: body
                  name: member
                  required: true
                  schema:
                    $ref: '#/definitions/MemberRequest'
            responses:
                "200":
                    description: The created member.
                    schema:
                        $ref: '#/definitions/MemberResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/MemberResponse'
            summary: Create a new member.
            tags:
                - members
    /internal/members/{id}:
        delete:
            description: Delete a member by ID.
            operationId: deleteMemberInternal
            parameters:
                - description: The ID of the member to delete.
                  in: path
                  name: id
                  required: true
                  type: integer
            responses:
                "200":
                    description: The deleted member.
                    schema:
                        $ref: '#/definitions/MemberResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/MemberResponse'
            summary: Delete a member.
            tags:
                - members
        put:
            description: Update a member with the provided information.
            operationId: updateMemberInternal
            parameters:
                - description: The ID of the member to update.
                  in: path
                  name: id
                  required: true
                  type: integer
                - description: The member to update.
                  in: body
                  name: member
                  required: true
                  schema:
                    $ref: '#/definitions/MemberRequest'
            responses:
                "200":
                    description: The updated member.
                    schema:
                        $ref: '#/definitions/MemberResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/MemberResponse'
            summary: Update a member.
            tags:
                - members
...
    /public/members:
        get:
            description: Get a list of all members in the system.
            operationId: getMembers
            responses:
                "200":
                    description: A list of members.
                    schema:
                        $ref: '#/definitions/MemberResponse'
                "400":
                    description: Bad request.
                    schema:
                        $ref: '#/definitions/MemberResponse'
            summary: Get a list of members.
            tags:
                - members
...
produces:
    - application/json
schemes:
    - http
swagger: "2.0"